% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_types_of_var.R
\name{convert_all_variables}
\alias{convert_all_variables}
\title{Convert type of variables}
\usage{
convert_all_variables(data, from, to)
}
\arguments{
\item{data}{dataset}

\item{from}{type of variable to be converted from}

\item{to}{type of variable to be converted to}
}
\value{
different type of variable
}
\description{
Convert a variable from one type to another type
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{nse-defuse}}
 \code{\link[dplyr]{mutate}}
}
