% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_tidy.R
\name{do_linear}
\alias{do_linear}
\title{Linear regression outputs}
\usage{
do_linear(a, b)
}
\arguments{
\item{a}{formula for linear model}

\item{b}{dataset}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Will run linear regression, then tidy the outputs
    and then glance
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[broom]{reexports}}
}
