% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda4num.R
\name{eda4num}
\alias{eda4num}
\title{Exploratory data analysis for numerical variables}
\usage{
eda4num(var_num, var_cat, dataset)
}
\arguments{
\item{var_num}{a numerical variable}

\item{var_cat}{a categorical variable}

\item{dataset}{the dataset}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Will generate boxplot and summary data
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}
 \code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}}
}
